+++ docs/REQUIREMNET_3.md
@@ -28,7 +28,7 @@
 - [x] Button의 Wrapper 코드 가독성 높여보기 (삼항 연산자 개선, 한줄 비우기 등)
 - [x] 코드 작성 중 depth가 2 깊이 이상 들어가지 않도록 한다.
 - [x] 파일의 라인 수가 200줄 이상이라면 코드의 흐름 파악이 용이하게 되도록 styled 코드를 분리한다.
-- [ ] onCardInfoValidateAndGoRegisterPage 함수를 개선해본다.
+- [x] onCardInfoValidateAndGoRegisterPage 함수를 개선해본다.
 - [ ] useFormInputs() 훅이 많은 input의 정보를 담는다면 작은 규모의 폼에서도 많은 인풋들의 정보를 불러오게 되어 좋지 않다고 생각한다. 현재는 제어 컴포넌트로 되어 있는데 비제어 컴포넌트로 고쳐서 훅의 재사용성을 높이고 사이트의 규모가 커져도 성능 이슈가 나지 않도록 해본다.
 - [ ] Context API는 전역으로만 사용하는 것이 아닌 props를 주는 또 다른 방법이다. 이러한 점을 고려해서 사용한다.
 - [ ] hook을 테스트하고, 스토리북에서 테스트하는 코드를 작성한다.
