--- src/pages/RegisterCard.tsx/index.ts
@@ -1,1 +1,0 @@
-export { default as RegisterCard } from './RegisterCard';
