+++ src/pages/RegisterCardPage.tsx/index.ts
@@ -1,0 +1,1 @@
+export { default as RegisterCardPage } from './RegisterCardPage';
